+++ src/pecha_framework/__init__.py
@@ -1,4 +1,87 @@
-from pecha_framework.pipeline import Document, Pipeline, component_registry
+from typing import List
+
+# Global component registry
+component_registry = {}
+
+
+class Pipe:
+    requires: List[str] = []
+
+    def __init_subclass__(cls, **kwargs):
+        """Automatically register subclasses in the component registry."""
+        super().__init_subclass__(**kwargs)
+        name = getattr(cls, "name", cls.__name__.lower())
+        component_registry[name] = cls
+        if not hasattr(cls, "requires"):
+            cls.requires = []
+
+
+class Document:
+    def __init__(self, text="", annotations=None, base_ann_mapping=None):
+        self.text = text
+        self.annotations = annotations or {}
+        self.base_ann_mapping = base_ann_mapping or []
+
+    def __repr__(self):
+        return f"Document(text={self.text[:50]!r}..., annotations={list(self.annotations.keys())})"
+
+    def get_attr(self, attr: str):
+        """Get the attribute of the document."""
+        attr_val = eval(f"self.{attr}")
+        assert isinstance(attr_val, str), f"Attribute '{attr}' is not a string."
+
+        return attr_val
+
+
+class Pipeline:
+    def __init__(self, components=None, **component_kwargs):
+        self.components = []
+        if components:
+            # check if the components requirements are met.
+            self.check_dependencies(components)
+            # Add the components to the pipeline
+            for name in components:
+                self.add_pipe(name, **component_kwargs.get(name, {}))
+
+    def check_dependencies(self, components):
+        """
+        Check if the component depending on other components are present before in the
+        pipeline.
+        """
+        for idx, component in enumerate(components):
+            component_cls = component_registry[component]
+            for required_component in component_cls.requires:
+                if required_component not in components[:idx]:
+                    raise ValueError(
+                        f"Component '{component}' requires '{required_component}'"
+                    )
+
+    def add_pipe(self, name, **kwargs):
+        """Add a component to the pipeline by name."""
+        if name not in component_registry:
+            raise ValueError(f"Component '{name}' not found in the registry.")
+
+        component_cls = component_registry[name]
+
+        # Check if the component is a function (callable) or a class
+        if callable(component_cls) and not isinstance(component_cls, type):
+            # If it's a function, we don't need to instantiate it
+            component = component_cls
+        else:
+            # If it's a class, instantiate it with the provided arguments
+            component = component_cls(**kwargs)
+
+        self.components.append((name, component))
+
+    def __call__(self, doc):
+        """Process the document through the pipeline."""
+        for name, component in self.components:
+            doc = component(doc)
+        return doc
+
+    def pipe_names(self):
+        """List the names of the components in the pipeline."""
+        return [name for name, _ in self.components]
 
 
 def pipe_component(name=None, requires=None):
