+++ tests/pipeline/pipeline_as_pipe/test_pipeline_as_pipe.py
@@ -1,6 +1,6 @@
 import pytest
 
-from pecha_framework.pipeline import Document, Pipe, Pipeline
+from pecha_framework import Document, Pipe, Pipeline
 
 
 class PlainTextParser(Pipe):
