+++ tests/stam_pipe/ann_text_mapping/simple/test_pipes_mapping.py
@@ -1,4 +1,4 @@
-from pecha_framework.pipeline import Document, Pipe, Pipeline
+from pecha_framework import Document, Pipe, Pipeline
 
 
 class FilterText(Pipe):
