--- tests/stam_pipe/stam_writer/output/I255A283F/base/F42F.txt
@@ -1,2 +1,0 @@
-Hello, world!
-This is a test.
