--- tests/stam_pipe/stam_writer/output/I5B364C43/base/ADE7.txt
@@ -1,2 +1,0 @@
-Hello, world!
-This is a test.
